Make github_release truly optional
While listed as an optional dependency in pyproject.toml, github_release
had to be installed in order to invoke AutoPub without a fatal error,
making the dependency, in practice, quite mandatory.

This restores the original optional intent. Now, the import only occurs
if the relevant function is called, and a helpful error message is
returned if the function's dependency is absent.